SpectralClusteringTools

Tools for Spectral Clustering.

A package for creating test data and providing functions to facilitate clustering analysis.

Package Features

Tools for generating synthetic test datasets for clustering experiments.

Functions for preprocessing data to prepare for spectral clustering.

Implementation of spectral clustering algorithms.

Utilities for visualizing clustering results and data distributions.

Getting Started

Installation

Install using the Julia's package manager. In the Julia REPL, type `] for entering the Pkg REPL mode and run

pkg> add "https://github.com/702ph/SpectralClusteringTools.jl"

type CTRL + C to back the Julia REPL. Equivalently with the Pkg API:

julia> import Pkg 
Julia> Pkg.add(url="https://github.com/702ph/SpectralClusteringTools.jl")

Example

using SpectralClusteringTools
using Plots
import PlotlyJS

# Use Plotly as the backend for Plots.jl
plotlyjs()

# A function for visualize given data as a 3D scatter plot grouped by labels.
function plot_data(data, labels, title="Visualization")
    scatter3d(data[:, 1], data[:, 2], data[:, 3],
              group=labels, title=title,
              xlabel="X", ylabel="Y", zlabel="Z", aspect_ratio=:auto,
              markersize=0.5)
end

# Generate test data
sphere_data, sphere_labels = make_sphere(3, 200)

# Visualize data
plot_data(sphere_data, sphere_labels, "Multiclass Spheres(Original)")

# Clustering 
#TODO: add or fixcode here
graph = epsilon_neighborhood_graph(sphere_data)
???? = next_fuction(graph)

clustered_labels = ????(???)

# Visualize
#TODO: add or fix code here
plot_data(sphere_data, clustered_labels, "Multiclass Spheres(Clustered)")

Documentation

Documentation for SpectralClusteringTools's interface.

SpectralClusteringTools.SpectralClusteringParams

SpectralClusteringTools.fully_connected_graph

SpectralClusteringTools.make_sphere

SpectralClusteringTools.timestwo

SpectralClusteringTools.SpectralClusteringParams — Type

Construct a similarity graph. W: weighted adjacency matrix(A Tutorial on Spectral Clustering P7)

● According to the reference, there are three ways to do it: (1) The ε-neighborhood graph (2) k-nearest neighbor graphs (3) The fully connected graph ● Parameters for Spectral clustering

SpectralClusteringTools.fully_connected_graph — Method

● Construct fully connected graph ● Connect all points with positive similarity with each other ● Weight all edge by sij ● Only useful if the similarity function itself models local neighbors， example: Gaussian

SpectralClusteringTools.make_sphere — Function

make_sphere(num_classes::Int, num_points_per_class::Int; adjust_scale::Bool=true)

Generates a dataset of points distributed on the surfaces of spheres。 Each sphere corresponds to a class, and points are normalized.

Arguments

num_classes: Number of spheres (classes) to generate.

num_points_per_class: Number of points per class.

adjust_scale: Adjusts the number of points.

Returns

A tuple of (normalized_points, labels), where normalized_points is a matrix of 3D points, and labels indicates class membership.

Example

points, labels = make_sphere(3, 1000, false)

SpectralClusteringTools.timestwo — Method

timestwo(x)

Multiply the input x by two.